Generate template builder
